Move binding createdBy tests
