+++ internal/broker/created_by.go
@@ -1,0 +1,10 @@
+package broker
+
+func (b *BindingContext) CreatedBy() string {
+	if b.Email != nil && *b.Email != "" {
+		return *b.Email
+	} else if b.Origin != nil && *b.Origin != "" {
+		return *b.Origin
+	}
+	return ""
+}
